Added --enable-opencl configuration option and library.
When --enable-opencl is specified at configure time, -lOpenCL is added to
the linked library list and runtime-cpu-detection is forced to on. The RTCD
linked -lOpenCL might be nice to get around to allow run-time detection of
OpenCL capabilities (through dlopen), but I'll try that later. For now I just
want to get basic OpenCL working in vpxdec.