Pick an appropriate global work queue size.
Previously, the maximum local work group size was being set as the global
work queue size, which was wrong... On this machine that meant that the CPU
was spawning 1024 threads to service a much smaller work queue. It also meant
that if the real work queue was > 1024 items, some of the work wouldn't get
done... Still working on trying to correct the output of the filter.