Added an intSize member to VP8_COMMON_CL.
This member tracks the allocated size in bytes of intData.
The first and second pass filters don't actually pass the same
size values around, which makes it pretty hard to compare things
and also causes some of the buffer reads to fail... It might just
be that the src_len param of the 2nd pass is calculated wrong, but
for now we cache the value.