Add working FILTER_OFFSET cases for 2nd pass filter.
Currently both the offset and buffer cases are about
the same speed on my machine.

Next, I will attempt to combine the two so that an
initial buffer of all of the filter data is copied
to the CL device and then a pointer and offset will
be passed in to each kernel invocation.