Added CL_ENSURE_BUF_SIZE macro.
This macro firstly makes sure that a CL buffer is allocated, and then
ensures that it is AT LEAST as big as what is needed. Then, if any data
has to be copied to initialize the buffer to the needed values, it does that.

This really cuts down on the amount of boiler-plate CL createBuffer, enqueueWrite,
and CL return code checking.