First attempt at dynamic loading of OpenCL library.
Uses a bunch of typedefs to declare the function prototypes and
then a large structure which holds reference to those functions.

dlopen() is used to populate the struct, and it's all tied together
by a loading function.

define's are used to swap out the actual cl* calls in code for
the real function pointers.

NOTE: I still need to dlclose() during encoder/decoder destruction,
and -lOpenCL is still being linked into the final binaries (and then ignored).

NOTE 2: Currently, EVERY function in the CL 1.1 spec is dynamically loaded...
This is overkill and might bite me later on (if a CL library that only
supports CL 1.0 is used.  Once the majority of macroblock decoding and
performance optimization is complete, it would be good to remove all
unused CL functions from dynamic_cl.[ch].