Moved clFinish() from filter_cl.c to reconinter.c.
The idea goes like this:
We don't need to clFinish (or enqueue a barrier) for every
call of the subpixel filters.  We can just keep enqueueing
more work for the CL library to do. As long as we make sure
that the CL threads are done (and output copied) BEFORE
we try to do anything on the CPU, the output is ok.

This means that if we have a loop which calls the subpixel
filters multiple times in a row, we are basically queuing up
a bunch of work for the GPU to do, but don't actually block
for each subpixel filter execution.

I have verified that the correct output is received from the
decoder, but so far the performance has been the same.  The major
downside to the current method is probably that the data
is still being copied to/from the GPU for each and every subpixel
filter call. Eventually, I need to find a way to just copy over
the entire set of macroblock/frame source data and do all operations
on the GPU (and move all CPU-only functions to kernels).