Fixed CL compiler warnings/errors for idct cl.
For the AMD compiler they were warnings but handled correctly ~50%
of the time (the other half it crashed). For NVidia compiler, it was
failing to compile every time.

As an aside, I've create a CL test compiler. It's basic, but given
a filename, it will load and test compile the CL source and spit
out the compiler output if it fails. This is a response to the
fact that I haven't managed to get my own VP8 CL compiler failure
output to work without crashing.