Add darwin10 x86 to build targets and get it working.
Needed to force x86-only for now. Eventually can add 64-bit x86.
Also had to add the appropriate -framework OpenCL library to
the compiler flags and force the dynamic CL option off.

As of now, my Macbook running Mac OS 10.6.5 is running the
opencl decoder with correct results.

Note that nasm doesn't work. You can only compile with yasm as
the assembler. This is a problem with upstream master as well
at this time.