Add some platform/device detection logic.
Trying to track down why using the GPU on my laptop is causing Bus Errors.
As far as I can tell, the Apple CL platform is the only one present and it
says that the GPU is available, but for some reason using it fails. Possibly
a driver re-install will help.