Start to re-work opencl initialization in prep for block memory refactor.
I'm planning on changing the Macroblock/Blockd memory allocation so that
instead of BLOCKD containing just a pointer to each of its bits of data, it
will have a base pointer and an offset. Then the function prototypes for
the SUBPIX_INVOKE, IDCT_INVOKE, RECON_INVOKE can all be changed to use the
base/offset mechanism. This is needed to prevent multiple copies of the same
basic data to gpu memory.