Loop through CL devices and check for byte-addressable-store extension.
Many of the kernels which have been written require this extension,
but I've been blindly assuming it was available until now... This was
made ok by the fact that I haven't found an implementation that doesn't
support it, but I was thinking that this was causing Mac GPU crashes...
unfortunately, it's supported, so I have to keep looking.

This code also checks ALL available devices on a system, and not just the
first device reported on a given platform. Should be a bit more robust
that way, and prefer GPU devices when available.

Note: I just realized that I need to break out of the platform enumeration
loop if a valid device is found... I'm not sure git will commit that change
right now, or it I need to do a follow-up commit.