Turns out that you ALWAYS need to check for byte_addressable_store.
Just because a platform supports the feature doesn't mean that
a device does.  Case in point: my Radeon 4770. I'm using Stream 2.3,
which supports CL 1.1, but the Radeon 4000 series doesn't support
the required functionality. I'd have to upgrade to a Radeon 5xxx/6xxx
or an Nvidia to get this system (desktop) to work.