Got a few of the macroblock dc/idct add y/uv functions working.
Moving in the direction of only copying the Block/Macroblock pointer
and varying arguments to the CL functions. This should eventually
lead to the ability to load/fetch the cl_mem contents at the
beginning and ending of a batch of commands.

Next: Need to remove cl_data.commands and cl_data.srcData/dstData.