Split off reconinter.* into reconinter_cl.*.
As a result, just do detection of CL in reconinter.* and call CL
functions if CL is initialized.

This should reduce the need to check cl_initialized all the time
as well, as we can safely assume that CL has been successfully
initialized in any of the reconinter/subpixel prediction functions.

Might need to add a return value to the functions, just in case
one of them fails and we need to fall back to the CPU path.