Initialize buffer_mem to null when CL not initialized.
If library built with OpenCL, this gives the client the ability to get
to a pointer for the buffer_mem instead of the raw frame data. It's probably
not too useful for my thesis project, but it could be used in the future by
someone who wants to blit the buffer contents directly to the screen, probably
after color space conversion.