Only send offsets_mem to GPU once per priority level and reduce argument count.
Reduce argument count to most of the C-based loop filter functions by creating
a struct containing the existing argument list. Still needs to filter out to
callers.

Also, change offsets_mem to be 5*num_blocks in size instead of 3*blocks. Now,
when doing a YUV pass followed by 2 additional Y passes, they are stored as
YUVYUVYUVYYYYYY (3 block example). This means that we can send only 1 buffer
to the GPU instead of sending the buffer 3 times (1 for YUV, 2 for Y passes).