Fixed the bugs that prevented multi-level filtering for small-ish groups.
The num_blocks was being set to the number for the starting
priority_level instead of the max of all levels.  Also, the num_blocks
in the kernel workers was not being calculated correctly. It was using
the maximum of the block count, not the correct count for the current
level.