AMD on Windows offers a better alternative to CL_MEM_ALLOC_HOST_PTR.
When using CL_MEM_ALLOC_HOST_PTR, the buffer is allocated in host
memory, and accessed on the GPU over the PCI-e bus.

The CL_MEM_USE_PERSISTENT_MEM_AMD option allows locating the buffer on
the GPU memory, which might speed up transfers to the GPU and accesses
from the GPU in kernels. It has a downside, which is that you need to
NOT use mapped memory to read the final buffer.